Assign labels to users

+++ app/controllers/labels_controller.rb
@@ -1,16 +1,20 @@
 class LabelsController < ApplicationController
   before_action :authenticate_user!
+  after_action :verify_authorized
 
   def index
     @labels = Label.all
+    authorize Label
   end
 
   def new
     @label = Label.new
+    authorize @label
   end
 
   def create
     @label = Label.new(secure_params)
+    authorize @label
 
     if @label.save
       flash[:notice] = "The label was created."
@@ -22,10 +26,12 @@ class LabelsController < ApplicationController
 
   def edit
     @label = Label.find(params[:id])
+    authorize @label
   end
 
   def update
     @label = Label.find(params[:id])
+    authorize @label
 
     if @label.update(secure_params)
       flash[:notice] = "The label was updated."
@@ -37,6 +43,7 @@ class LabelsController < ApplicationController
 
   def destroy
     label = Label.find(params[:id])
+    authorize label
 
     if label.destroy
       flash[:notice] = "The label was destroyed."
